+++ src/NexNet.IntegrationTests/Pipes/NexusServerTests_NexusDuplexPipe.cs
@@ -598,50 +598,4 @@ internal class NexusServerTests_NexusDuplexPipe : BasePipeTests
         await AssertThrows<InvalidOperationException>(() =>
             sNexus.Context.Clients.Caller.ClientTaskValueWithDuplexPipe(pipe).Timeout(1)).Timeout(1);
     }
-    
-    [TestCase(Type.Quic)]
-    [TestCase(Type.Uds)]
-    [TestCase(Type.Tcp)]
-    [TestCase(Type.TcpTls)]
-    [TestCase(Type.WebSocket)]
-    [TestCase(Type.HttpSocket)]
-    public async Task Server_ChannelReader_IAsyncEnumerable(Type type)
-    {
-        var (_, sNexus, _, cNexus, _) = await Setup(type);
-        var tcs = new TaskCompletionSource();
-        cNexus.ClientTaskValueWithDuplexPipeEvent = async (nexus, pipe) =>
-        {
-            var writer = await pipe.GetUnmanagedChannelWriter<int>();
-            int counter = 0;
-            await foreach (var message in await pipe.GetChannelReader<ComplexMessage>())
-            {
-                await writer.WriteAsync(counter++);
-            }
-        };
-
-        var pipe = sNexus.Context.CreatePipe();
-        await sNexus.Context.Clients.Caller.ClientTaskValueWithDuplexPipe(pipe).Timeout(1);
-        await pipe.ReadyTask.Timeout(1);
-        int counter = 0;
-
-        _ = Task.Run(async () =>
-        {
-            await foreach (var message in await pipe.GetUnmanagedChannelReader<int>())
-            {
-                Interlocked.Increment(ref counter);
-                if(counter == 10)
-                    tcs.SetResult();
-                
-            }
-        });
-        
-        var writer = await pipe.GetChannelWriter<ComplexMessage>();
-        for (int i = 0; i < 10; i++)
-        {
-            await writer.WriteAsync(ComplexMessage.Random());
-            await Task.Delay(1);
-        }
-        await writer.CompleteAsync().ConfigureAwait(false);
-        await tcs.Task.Timeout(1);
-    }
 }
